+++ CHANGES.rst
@@ -11,6 +11,7 @@ Fixes
 New Features
 ~~~~~~~~~~~~
 - Add ``filter_func`` superslab filtering to ``CompaSOHaloCatalog`` [#16]
+- Add pack9 reader [#25]
 
 Changes
 ~~~~~~~